requirements.txt
discord.py
discord-ext-slash
aiofiles